# ⚡️ Speed up function `_mask_reduce_batched` by 19%