### 📄 19% (0.19x) speedup for ***`_mask_reduce_batched` in `lightly/models/utils.py`***

⏱️ Runtime :   **`3.13 milliseconds`**  **→** **`2.64 milliseconds`** (best of `47` runs)
### 📝 Explanation and details

Here are several changes to improve performance, especially for large tensors.

- **Remove unnecessary `torch.arange` and `num_cls` manipulation.** You already receive `num_cls` as an int, and `torch.arange(num_cls, ...)` isn't used.  
- **Combine view/expand in-place and avoid allocating new tensors when not needed.**
- **If `output` does not require gradient, use `.requires_grad_(False)` so PyTorch can optimize memory in some settings (optional if strict, but included).**
- **Move `mask` expansion to a more memory-efficient `repeat_interleave`, which avoids a full broadcast (though for proper `scatter_reduce_` index use, the original expansion is already nearly minimal).**
- **Elide intermediates where possible.**
- **Minimize view operations.**

Here's the faster, memory-optimized rewrite.



**Summary of optimizations:**
- Removed extra `torch.arange` and `cls` unused allocation.
- Combined mask expansion with view efficiently using as few copies as possible.
- Used existing input tensors without superfluous temporary arrays.
- Kept the shape manipulations efficient.

The mathematical result is identical. The most critical time- and memory-saving fix is skipping the unnecessary creation/use of `torch.arange` and working with the mask/viewing in as few steps as possible.


✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | ✅ **42 Passed** |
| 🌀 Generated Regression Tests | ✅ **32 Passed** |
| ⏪ Replay Tests | ✅ **5 Passed** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>⚙️ Existing Unit Tests Details</summary>

```python
- models/test_ModelUtils.py
- test_tracer_py__replay_test_3.py
```

</details>

<details>
<summary>🌀 Generated Regression Tests Details</summary>

```python
from __future__ import annotations

# imports
import pytest  # used for our unit tests
import torch
from lightly.models.utils import _mask_reduce_batched
from torch import Tensor

# unit tests

# 1. Basic Test Cases

def test_basic_single_batch_single_channel_single_class():
    # Single batch, single channel, single class
    source = torch.tensor([[[[1.0]]]])  # shape (1, 1, 1, 1)
    mask = torch.tensor([[[0]]])        # shape (1, 1, 1)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=1); result = codeflash_output # 94.7μs -> 82.1μs

def test_basic_two_classes_mean():
    # Two classes, mean reduction
    source = torch.tensor([[[[1.0, 3.0], [5.0, 7.0]]]])  # (1,1,2,2)
    mask = torch.tensor([[[0, 1], [0, 1]]])              # (1,2,2)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="mean"); result = codeflash_output # 74.7μs -> 55.2μs
    # class 0: [1.0,5.0] mean=3.0; class 1: [3.0,7.0] mean=5.0
    expected = torch.tensor([[[3.0, 5.0]]])

def test_basic_multi_batch_multi_channel():
    # Two batches, two channels, three classes
    source = torch.tensor([
        [
            [[1, 2], [3, 4]],  # channel 0
            [[5, 6], [7, 8]],  # channel 1
        ],
        [
            [[9, 10], [11, 12]],
            [[13, 14], [15, 16]],
        ]
    ], dtype=torch.float32)  # (2,2,2,2)
    mask = torch.tensor([
        [[0, 1], [2, 1]],
        [[1, 2], [0, 2]],
    ])  # (2,2,2)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=3, reduce="mean"); result = codeflash_output # 74.3μs -> 59.2μs
    # Manually compute expected values for each batch, channel, class
    # batch 0, channel 0: class 0:1, class 1:2,4, class 2:3
    # batch 0, channel 0: [1, (2+4)/2=3, 3]
    # batch 0, channel 1: [5, (6+8)/2=7, 7]
    # batch 1, channel 0: [11, 9, (10+12)/2=11]
    # batch 1, channel 1: [15, 13, (14+16)/2=15]
    expected = torch.tensor([
        [[1, 3, 3],
         [5, 7, 7]],
        [[11, 9, 11],
         [15, 13, 15]]
    ], dtype=torch.float32)

def test_basic_sum_reduction():
    # Test sum reduction
    source = torch.tensor([[[[1, 2], [3, 4]]]], dtype=torch.float32)  # (1,1,2,2)
    mask = torch.tensor([[[0, 1], [0, 1]]])
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="sum"); result = codeflash_output # 51.7μs -> 34.2μs
    # class 0: 1+3=4, class 1: 2+4=6
    expected = torch.tensor([[[4, 6]]], dtype=torch.float32)

def test_basic_max_reduction():
    # Test max reduction
    source = torch.tensor([[[[1, 5], [3, 2]]]], dtype=torch.float32)  # (1,1,2,2)
    mask = torch.tensor([[[0, 1], [0, 1]]])
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="amax"); result = codeflash_output # 52.2μs -> 36.3μs
    # class 0: max(1,3)=3, class 1: max(5,2)=5
    expected = torch.tensor([[[3, 5]]], dtype=torch.float32)

# 2. Edge Test Cases

def test_edge_empty_class():
    # Some classes have no corresponding pixels in mask
    source = torch.tensor([[[[1, 2], [3, 4]]]], dtype=torch.float32)  # (1,1,2,2)
    mask = torch.tensor([[[1, 1], [1, 1]]])
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=3, reduce="mean"); result = codeflash_output # 72.3μs -> 57.7μs
    # class 0: empty -> should be 0, class 1: mean(1,2,3,4)=2.5, class 2: empty->0
    expected = torch.tensor([[[0.0, 2.5, 0.0]]], dtype=torch.float32)






def test_edge_zero_height_width():
    # H=0 or W=0 should return shape (B, C, N)
    source = torch.randn(1, 2, 0, 3)
    mask = torch.randint(0, 2, (1, 0, 3))
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="sum"); result = codeflash_output # 18.0μs -> 13.5μs

    source = torch.randn(1, 2, 3, 0)
    mask = torch.randint(0, 2, (1, 3, 0))
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="sum"); result = codeflash_output # 18.0μs -> 13.5μs


def test_edge_non_contiguous_input():
    # Non-contiguous input tensors
    source = torch.arange(24, dtype=torch.float32).reshape(2, 3, 2, 2).transpose(0, 1)
    mask = torch.tensor([
        [[0, 1], [1, 0]],
        [[1, 0], [0, 1]],
        [[0, 1], [1, 0]]
    ], dtype=torch.long)  # shape (3,2,2)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=2, reduce="sum"); result = codeflash_output # 69.7μs -> 65.5μs

# 3. Large Scale Test Cases

def test_large_scale_batch_and_classes():
    # Large batch size and number of classes, but under 100MB
    B, C, H, W, N = 8, 4, 8, 8, 16  # (8*4*8*8*4 bytes = 8KB)
    source = torch.arange(B*C*H*W, dtype=torch.float32).reshape(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=N, reduce="sum"); result = codeflash_output # 54.6μs -> 43.9μs

def test_large_scale_many_channels():
    # Large number of channels
    B, C, H, W, N = 2, 256, 4, 4, 5  # (2*256*4*4*4 bytes = 32KB)
    source = torch.randn(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=N, reduce="mean"); result = codeflash_output # 148μs -> 132μs

def test_large_scale_randomized():
    # Randomized test with moderate size
    torch.manual_seed(42)
    B, C, H, W, N = 4, 3, 16, 16, 10
    source = torch.randn(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=N, reduce="mean"); result = codeflash_output # 94.2μs -> 78.3μs
    # For each class, the mean should be between min and max of source
    min_val = source.min().item()
    max_val = source.max().item()

def test_large_scale_performance():
    # This test is not a strict performance test, but ensures that the function runs in reasonable time
    B, C, H, W, N = 8, 8, 32, 32, 32  # (8*8*32*32*4 bytes = 256KB)
    source = torch.randn(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, num_cls=N, reduce="sum"); result = codeflash_output # 239μs -> 210μs

# 4. Error Handling

def test_invalid_reduce_string():
    # Should raise RuntimeError for unsupported reduce operation
    source = torch.randn(1, 1, 2, 2)
    mask = torch.zeros(1, 2, 2, dtype=torch.long)
    with pytest.raises(RuntimeError):
        _mask_reduce_batched(source, mask, num_cls=1, reduce="invalid_reduce")

def test_incompatible_mask_shape():
    # Mask shape incompatible with source
    source = torch.randn(1, 1, 2, 2)
    mask = torch.zeros(1, 3, 2, dtype=torch.long)
    with pytest.raises(RuntimeError):
        _mask_reduce_batched(source, mask, num_cls=1, reduce="sum")

def test_incompatible_source_mask_dtype():
    # Mask dtype not long/int
    source = torch.randn(1, 1, 2, 2)
    mask = torch.zeros(1, 2, 2, dtype=torch.float32)
    with pytest.raises(RuntimeError):
        _mask_reduce_batched(source, mask, num_cls=1, reduce="sum")
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.

from __future__ import annotations

# imports
import pytest  # used for our unit tests
import torch
from lightly.models.utils import _mask_reduce_batched
from torch import Tensor

# unit tests

# 1. Basic Test Cases

def test_single_batch_single_channel_single_class():
    # Single batch, single channel, single class
    source = torch.tensor([[[[1.0]]]])  # shape (1, 1, 1, 1)
    mask = torch.tensor([[[0]]])        # shape (1, 1, 1)
    num_cls = 1
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="mean"); result = codeflash_output # 99.2μs -> 81.0μs

def test_single_batch_single_channel_two_classes():
    # Single batch, single channel, two classes
    source = torch.tensor([[[[1.0, 2.0]]]])  # shape (1, 1, 1, 2)
    mask = torch.tensor([[[0, 1]]])          # shape (1, 1, 2)
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="mean"); result = codeflash_output # 77.8μs -> 57.1μs

def test_single_batch_multi_channel_multi_class_sum():
    # Single batch, two channels, two classes, reduce="sum"
    source = torch.tensor([[
        [[1.0, 2.0], [3.0, 4.0]],  # channel 0
        [[5.0, 6.0], [7.0, 8.0]],  # channel 1
    ]])  # shape (1, 2, 2, 2)
    mask = torch.tensor([[
        [0, 1],
        [1, 0],
    ]])  # shape (1, 2, 2)
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="sum"); result = codeflash_output # 50.8μs -> 35.0μs

def test_single_batch_multi_channel_multi_class_mean():
    # Single batch, two channels, two classes, reduce="mean"
    source = torch.tensor([[
        [[1.0, 2.0], [3.0, 4.0]],  # channel 0
        [[5.0, 6.0], [7.0, 8.0]],  # channel 1
    ]])  # shape (1, 2, 2, 2)
    mask = torch.tensor([[
        [0, 1],
        [1, 0],
    ]])  # shape (1, 2, 2)
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="mean"); result = codeflash_output # 85.0μs -> 64.0μs

def test_multi_batch_multi_channel_multi_class():
    # Two batches, two channels, three classes
    source = torch.tensor([
        [  # batch 0
            [[1.0, 2.0], [3.0, 4.0]],  # channel 0
            [[5.0, 6.0], [7.0, 8.0]],  # channel 1
        ],
        [  # batch 1
            [[9.0, 10.0], [11.0, 12.0]],
            [[13.0, 14.0], [15.0, 16.0]],
        ]
    ])  # shape (2, 2, 2, 2)
    mask = torch.tensor([
        [[0, 1], [1, 2]],  # batch 0
        [[0, 1], [1, 2]],  # batch 1
    ])  # shape (2, 2, 2)
    num_cls = 3
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="sum"); result = codeflash_output # 51.9μs -> 36.4μs

# 2. Edge Test Cases

def test_mask_with_unseen_class():
    # Mask contains a class index not present in mask
    source = torch.ones((1, 1, 2, 2))
    mask = torch.zeros((1, 2, 2), dtype=torch.long)
    num_cls = 3  # classes 0, 1, 2; only class 0 present
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="sum"); result = codeflash_output # 43.1μs -> 33.2μs



def test_single_pixel_per_class():
    # Each class appears exactly once
    source = torch.tensor([[[[1.0, 2.0], [3.0, 4.0]]]])
    mask = torch.tensor([[[0, 1], [2, 3]]])
    num_cls = 4
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="mean"); result = codeflash_output # 113μs -> 110μs
    # Each class should get exactly one value
    for i in range(4):
        pass

def test_negative_mask_indices():
    # Negative indices in mask should raise an error
    source = torch.ones((1, 1, 2, 2))
    mask = torch.tensor([[[0, -1], [1, 0]]])
    num_cls = 2
    with pytest.raises(RuntimeError):
        _mask_reduce_batched(source, mask, num_cls, reduce="sum")



def test_reduce_max():
    # Test with reduce="amax"
    source = torch.tensor([[[[1.0, 2.0], [3.0, 2.0]]]])
    mask = torch.tensor([[[0, 1], [1, 0]]])
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="amax"); result = codeflash_output # 85.0μs -> 80.4μs

def test_reduce_min():
    # Test with reduce="amin"
    source = torch.tensor([[[[1.0, 2.0], [3.0, 2.0]]]])
    mask = torch.tensor([[[0, 1], [1, 0]]])
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="amin"); result = codeflash_output # 48.5μs -> 37.5μs

def test_reduce_prod():
    # Test with reduce="prod"
    source = torch.tensor([[[[2.0, 3.0], [4.0, 5.0]]]])
    mask = torch.tensor([[[0, 1], [1, 0]]])
    num_cls = 2
    codeflash_output = _mask_reduce_batched(source, mask, num_cls, reduce="prod"); result = codeflash_output # 47.1μs -> 37.2μs

# 3. Large Scale Test Cases

def test_large_batch_and_classes():
    # Test with large batch and many classes, but <100MB
    B, C, H, W, N = 16, 4, 8, 8, 10  # ~32KB
    source = torch.arange(B*C*H*W, dtype=torch.float32).reshape(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, N, reduce="sum"); result = codeflash_output # 66.0μs -> 53.9μs

def test_large_image():
    # Test with a single large image
    B, C, H, W, N = 1, 2, 64, 64, 8  # ~64*64*2*4 = 32KB
    source = torch.ones((B, C, H, W))
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, N, reduce="mean"); result = codeflash_output # 125μs -> 115μs

def test_large_number_of_classes_sparse():
    # Large number of classes, but only a few present in mask
    B, C, H, W, N = 2, 2, 10, 10, 500  # ~4KB
    source = torch.ones((B, C, H, W))
    mask = torch.zeros((B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, N, reduce="sum"); result = codeflash_output # 55.3μs -> 37.1μs

def test_large_randomized_input():
    # Random large input, check for runtime and shape correctness
    B, C, H, W, N = 8, 3, 32, 32, 20  # ~98KB
    torch.manual_seed(42)
    source = torch.randn(B, C, H, W)
    mask = torch.randint(0, N, (B, H, W), dtype=torch.long)
    codeflash_output = _mask_reduce_batched(source, mask, N, reduce="mean"); result = codeflash_output # 199μs -> 180μs

def test_large_all_classes_present():
    # All classes present in mask
    B, C, H, W, N = 4, 2, 16, 16, 16  # ~8KB
    source = torch.arange(B*C*H*W, dtype=torch.float32).reshape(B, C, H, W)
    mask = torch.arange(H*W).reshape(H, W) % N
    mask = mask.unsqueeze(0).expand(B, -1, -1)
    codeflash_output = _mask_reduce_batched(source, mask, N, reduce="sum"); result = codeflash_output # 48.1μs -> 39.0μs
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>


To edit these changes `git checkout codeflash/optimize-_mask_reduce_batched-mc0q5mek` and push.

[![Codeflash](https://img.shields.io/badge/Optimized%20with-Codeflash-yellow?style=flat&color=%23ffc428&logo=data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDgwIiBoZWlnaHQ9ImF1dG8iIHZpZXdCb3g9IjAgMCA0ODAgMjgwIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTI4Ni43IDAuMzc4NDE4SDIwMS43NTFMNTAuOTAxIDE0OC45MTFIMTM1Ljg1MUwwLjk2MDkzOCAyODEuOTk5SDk1LjQzNTJMMjgyLjMyNCA4OS45NjE2SDE5Ni4zNDVMMjg2LjcgMC4zNzg0MThaIiBmaWxsPSIjRkZDMDQzIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzExLjYwNyAwLjM3ODkwNkwyNTguNTc4IDU0Ljk1MjZIMzc5LjU2N0w0MzIuMzM5IDAuMzc4OTA2SDMxMS42MDdaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzA5LjU0NyA4OS45NjAxTDI1Ni41MTggMTQ0LjI3NkgzNzcuNTA2TDQzMC4wMjEgODkuNzAyNkgzMDkuNTQ3Vjg5Ljk2MDFaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMjQyLjg3MyAxNjQuNjZMMTg5Ljg0NCAyMTkuMjM0SDMxMC44MzNMMzYzLjM0NyAxNjQuNjZIMjQyLjg3M1oiIGZpbGw9IiMwQjBBMEEiLz4KPC9zdmc+Cg==)](https://codeflash.ai)